Refactor code review results
