+++ lib/config/definitions.js
@@ -1377,9 +1377,6 @@ const options = [
     default: {
       enabled: true,
       fileMatch: ['(^|/)gradle/wrapper/gradle-wrapper.properties$'],
-      typeStrategy: 'auto',
-      digestsStrategy: 'true',
-      serviceUrl: 'https://services.gradle.org/versions/all',
     },
     mergeable: true,
     cli: false,
