+++ lib/datasource/gradle-version/index.js
@@ -5,20 +5,11 @@ module.exports = {
   getPkgReleases,
 };
 
+const GradleVersionsServiceUrl = 'https://services.gradle.org/versions/all';
+
 async function getPkgReleases(purl, config = {}) {
   try {
-    const { typeStrategy, digestsStrategy, serviceUrl } = config;
-    let { type, digests } = config;
-
-    if (typeStrategy !== 'auto') {
-      type = typeStrategy;
-    }
-
-    if (digestsStrategy !== 'auto') {
-      digests = digestsStrategy;
-    }
-
-    const response = await got(serviceUrl, {
+    const response = await got(GradleVersionsServiceUrl, {
       json: true,
     });
     const releases = await Promise.all(
@@ -34,20 +25,7 @@ async function getPkgReleases(purl, config = {}) {
             version: coerce(release.version).toString(),
             downloadUrl: release.downloadUrl,
             checksumUrl: release.checksumUrl,
-            digests,
           };
-
-          if (type === 'all') {
-            entry.downloadUrl = replaceType(entry.downloadUrl);
-            entry.checksumUrl = replaceType(entry.checksumUrl);
-          }
-
-          if (digests === 'true') {
-            const checksum = await getChecksum(entry.checksumUrl);
-            if (checksum) {
-              entry.checksum = checksum;
-            }
-          }
 
           return entry;
         })
@@ -66,24 +44,5 @@ async function getPkgReleases(purl, config = {}) {
       logger.warn({ err }, 'Gradle release lookup failure: Unknown error');
     }
     return null;
-  }
-}
-
-function replaceType(url) {
-  return url.replace('bin', 'all');
-}
-
-async function getChecksum(url) {
-  try {
-    const response = await got(url, {});
-    return response.body;
-  } catch (err) {
-    if (err.statusCode === 404 || err.code === 'ENOTFOUND') {
-      logger.info('Gradle checksum lookup failure: not found');
-      logger.debug({ err });
-    } else {
-      logger.warn({ err }, 'Gradle checksum lookup failure: Unknown error');
-    }
-    throw err;
   }
 }
