+++ lib/datasource/index.js
@@ -7,7 +7,7 @@ const docker = require('./docker');
 const github = require('./github');
 const gitlab = require('./gitlab');
 const go = require('./go');
-const gradle = require('./gradle');
+const gradleVersion = require('./gradle-version');
 const maven = require('./maven');
 const npm = require('./npm');
 const nuget = require('./nuget');
@@ -22,7 +22,7 @@ const datasources = {
   github,
   gitlab,
   go,
-  gradle,
+  gradleVersion,
   maven,
   npm,
   nuget,
