+++ lib/manager/gradle-wrapper/extract.js
@@ -1,3 +1,5 @@
+const { coerce } = require('semver');
+
 module.exports = {
   extractPackageFile,
 };
@@ -9,26 +11,25 @@ function extractPackageFile(fileContent) {
   let lineNumber = 0;
   for (const line of lines) {
     const match = line.match(
-      /^distributionUrl=https\\:\/\/services.gradle.org\/distributions\/gradle-((\d|\.)+)-(bin|all)\.zip\s*$/
+      /^distributionUrl=.*-((\d|\.)+)-(bin|all)\.zip\s*$/
     );
     if (match) {
       const dependency = {
-        purl: 'pkg:gradle',
+        purl: 'pkg:gradleVersion',
         depType: 'gradle-wrapper',
         depName: 'gradle',
-        currentValue: match[1],
+        currentValue: coerce(match[1]).toString(),
         lineNumber,
         versionScheme: 'semver',
         /* gradle specifics ... */
-        type: match[3],
+        gradleWrapperType: match[3],
       };
 
       let shaLineNumber = 0;
       for (const shaLine of lines) {
         const shaMatch = shaLine.match(/^distributionSha256Sum=((\w){64}).*$/);
         if (shaMatch) {
-          dependency.shaCurrentValue = shaMatch[1];
-          dependency.shaLineNumber = shaLineNumber;
+          dependency.checksumLineNumber = shaLineNumber;
           break;
         }
         shaLineNumber += 1;
