+++ lib/manager/gradle-wrapper/update.js
@@ -1,32 +1,56 @@
+const got = require('got');
+
 module.exports = {
   updateDependency,
 };
 
-function updateDependency(fileContent, upgrade) {
+async function updateDependency(fileContent, upgrade) {
   try {
     logger.debug(upgrade, 'gradle-wrapper.updateDependency()');
     const lines = fileContent.split('\n');
 
+    if (upgrade.gradleWrapperType === 'all') {
+      upgrade.downloadUrl = replaceType(upgrade.downloadUrl);
+      upgrade.checksumUrl = replaceType(upgrade.checksumUrl);
+    }
+
     const url = upgrade.downloadUrl.replace(':', '\\:');
+    const checksum = await getChecksum(upgrade.checksumUrl);
+
     lines[upgrade.lineNumber] = `distributionUrl=${url}`;
 
-    if (upgrade.digests === 'true' && upgrade.shaLineNumber) {
-      lines[upgrade.shaLineNumber] = `distributionSha256Sum=${
-        upgrade.checksum
-      }`;
-    } else if (upgrade.digests === 'true') {
+    if (upgrade.checksumLineNumber) {
+      lines[upgrade.checksumLineNumber] = `distributionSha256Sum=${checksum}`;
+    } else {
       lines.splice(
         upgrade.lineNumber + 1,
         0,
-        `distributionSha256Sum=${upgrade.checksum}`
+        `distributionSha256Sum=${checksum}`
       );
-    } else if (upgrade.digests !== 'true' && upgrade.shaLineNumber) {
-      lines.splice(upgrade.shaLineNumber, 1);
     }
 
     return lines.join('\n');
   } catch (err) {
     logger.info({ err }, 'Error setting new Gradle Wrapper release value');
     return null;
+  }
+}
+
+function replaceType(url) {
+  return url.replace('bin', 'all');
+}
+
+async function getChecksum(url) {
+  try {
+    const response = await got(url, {});
+    return response.body;
+  } catch (err) {
+    if (err.statusCode === 404 || err.code === 'ENOTFOUND') {
+      logger.info('Gradle checksum lookup failure: not found');
+      logger.debug({ err });
+    } else {
+      logger.warn({ err }, 'Gradle checksum lookup failure: Unknown error');
+    }
+    throw err;
   }
 }
